producer1/requirements.txt
requests
pymongo
kafka-python
requests